services/event-ingest-stream/requirements.txt
fastapi
uvicorn[standard]
redis[asyncio]
asyncpg
pydantic
python-jose[cryptography]
prometheus-fastapi-instrumentator
python--dotenv